Backport core afform fix
I discovered the new search kit 'bug' is because
we are actually using an old version of afform due
to it being in our extensions folder. When I tried
to switch over I hit a bug fixed by this patch. I
discussed the patch with Tim & Coleman and it is now
merged upstream

https: //github.com/civicrm/civicrm-core/pull/19670

Note this should be a no-change-patch at this stage as we are
not actually using this code - but it's a
prerequisite for my patch that fixes that.

Bug: T275728
Change-Id: I6dfaf09abd3617244fedf41c464420f826c20232